refactor module to decouple and do 2.5 exercice
