+++ m2/courseinfo/src/App.js
@@ -1,28 +1,5 @@
 import {appContent} from './data/course'
-import {
-  Header, 
-  SubHeader,
-  Content, 
-  Footer,
-  getTotal
-} from './components/allComponents'
-
-
-
-const Course = ({course}) => {
-
-    const total = getTotal(course)
-
-    return (
-      <div id="outer">
-          <SubHeader name={course['name']} id={course['id']} />
-          <Content lines = {course['parts']} />
-          <Footer line ={total} />
-      </div>
-  )
-
-}
-
+import {Header} from './components/allComponents'
 
 const App = () => {
 
@@ -37,6 +14,5 @@ const App = () => {
         }</>
     ]
 }
-
 
 export default App;
